requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
pydantic==2.8.2
slowapi==0.1.9

asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
psycopg2-binary
greenlet
